[Feat] : log 파일 정리(#77)
