--- src/main/java/com/potato/balbambalbam/log/TestController.java
@@ -1,12 +1,0 @@
-package com.potato.balbambalbam.log;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class TestController {
-    @GetMapping("/test-exception")
-    public void testException() {
-        throw new RuntimeException("Test exception");
-    }
-}
