+++ src/main/java/com/potato/balbambalbam/log/controller/TestController.java
@@ -1,0 +1,12 @@
+package com.potato.balbambalbam.log.controller;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class TestController {
+    @GetMapping("/test-exception")
+    public void testException() {
+        throw new RuntimeException("Test exception");
+    }
+}
